session fix
I was finding that when a JWT was not processed, the session['user']
set to current_user, which might itself return session['user'] from a
previous JWT. If the session contained an openstruct, serializing it
would result in { "table" => { "email" => "...", "groups" => [...] }}.
Each time this gets converted to an OpenStruct, it gains another level
of nesting under "table".
So. Use a version of OpenStruct that doesn't nest its fields under
"table" when serialized.
Also, clear the session at the start of check_credentials
so that sessions from JWTs read once won't be reused in subsequent requests.